ima: LSM-based policy for "floor" files
The policy enables appraisal for all files with the _ (aka floor)
label. Typically these are read-only files, so the policy does not
interfere (much) with booting a system.

However, such a policy most likely is not secure: an off-line attacker
merely needs to label a file as "System" instead of "_" and then can
get a modified file used by the running system because the file
content would no longer be checked.

LSM-based IMA policies depend on enabling the audit sub-system,
which may be undesirable (performance impact?).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>