e2fsprogs: replace with meta-selinux version
The bbappends from meta-selinux (revision 2188d5b09b8f79cce93589) add
a different set of patches. The ones we had before had known issues
which cause image corruption when fsck.ext4 runs, in particular (or
only?) when system time is not correct (see
https://bugzilla.yoctoproject.org/show_bug.cgi?id=8544).

There is a pending request to move the meta-selinux code to OE-core
(https://bugzilla.yoctoproject.org/show_bug.cgi?id=8622) but as it is
unclear when that will happen, it makes sense to update it here instead
of waiting further.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>