smack config: remove System::Log
The setup of System::Log already was incomplete in Tizen: the
directory "transmute" flag had no effect because the actual Smack
rules did not include transmute.

Having the System::Log is questionable, with no clear use cases
obvious for it. Therefore instead of fixing the setup, the special
domain gets removed entirely.

Fixes: IOTOS-1503

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>